# LineCrossingWaterbodyCheck performance improvements