--- a/config/configuration.json
+++ b/config/configuration.json
@@ -474,7 +474,9 @@
       "blurb": "Edit features overlapping the waterbody so they either validly overlap or do not overlap.",
       "instruction": "Open your favorite editor and edit the features overlapping the waterbody so they either validly overlap or do not overlap.",
       "difficulty": "EASY"
-    }
+    },
+    "shapepoints.max": 1,
+    "shapepoints.max": 5000
   },
   "LongSegmentCheck": {
     "length.minimum.kilometers": 10.0,