--- a/src/main/java/org/openstreetmap/atlas/checks/validation/intersections/LineCrossingWaterBodyCheck.java
+++ b/src/main/java/org/openstreetmap/atlas/checks/validation/intersections/LineCrossingWaterBodyCheck.java
@@ -426,7 +426,7 @@ public class LineCrossingWaterBodyCheck extends BaseCheck<Long>
                     .intersections(intersectingFeature);
             if (intersectionLocations.isEmpty())
             {
-                return Set.of();
+                return Set.of(Tuple.createTuple((Polygon) waterbody, Set.of()));
             }
             return Set.of(Tuple.createTuple((Polygon) waterbody, intersectionLocations));
         }