@danielduhh 2020-07-14
Hey @seancoulter! Do you have any before/after metrics for some of the problematic shards?